+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## TO BE RELEASED
 
 * insert missing `install-cwlogs` recipe in zadara cluster config template
+* add `num_workers=n` option to `stack:instances:start` to allow only starting *n* workers.
 
 ## 1.16.1 - 12/13/2017